refactor: move preview in a new util
